backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
python-dotenv
google-genai
pydantic
